Model chanages, admin page changes, kitchen changes
For models, the old OrderItem to food relationship didn't allow for quantities so I added that. I also added some restrictions on the order types and food item types. For kitchen I edited the timer a little bit. And for the admin page I installed a plugin to allow for recursive inline object creation. That is, from the order page, I can specify the order items, and those order items' food items